+++ spec/client/realm_client_spec.rb
@@ -96,4 +96,30 @@ RSpec.describe KeycloakAdmin::RealmClient do
       expect(@realm_client.delete).to be_truthy
     end
   end
+
+  describe "#update" do
+    let(:realm_name) { "valid-realm" }
+
+    before(:each) do
+      @realm_client = KeycloakAdmin.realm(realm_name)
+
+      stub_token_client
+    end
+
+    it "updates realm" do
+      expect_any_instance_of(RestClient::Resource).to receive(:put).with('{"smtpServer":{"host":"test_host"}}', anything)
+      @realm_client.update({ smtpServer: { host: 'test_host' } })
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm", rest_client_options).and_call_original
+
+      expect_any_instance_of(RestClient::Resource).to receive(:put).with('{"smtpServer":{"host":"test_host"}}', anything)
+      @realm_client.update({ smtpServer: { host: 'test_host' } })
+    end
+  end
 end